+++ examples/tuned_catboost.py
@@ -1,14 +1,14 @@
 """
-Reported (reproduced) results of Tuned XGBoost on TabularBenchmark of
+Reported (reproduced) results of Tuned CatBoost on TabularBenchmark of
 the Trompt paper: https://arxiv.org/abs/2305.18446
-electricity (A4): 88.52 (91.09)
-eye_movements (A5): 66.57 (64.21)
-MagicTelescope (B2): 86.05 (86.50)
-bank-marketing (B4): 80.34 (80.41)
-california (B5): 90.12 (89.71)
-credit (B7): 77.26 (77.4)
-pol (B14): 98.09 (97.5)
-jannis (mathcal B4): 79.67 (77.81)
+electricity (A4): 87.73 (88.09)
+eye_movements (A5): 66.84 (64.27)
+MagicTelescope (B2): 85.92 (87.18)
+bank-marketing (B4): 80.39 (80.50)
+california (B5): 90.32 (87.56)
+credit (B7): 77.59 (77.29)
+pol (B14): 98.49 (98.21)
+jannis (mathcal B4): 79.89 (78.96)
 """
 import argparse
 import os.path as osp
@@ -19,10 +19,10 @@ import torch
 
 from torch_frame import TaskType
 from torch_frame.datasets import TabularBenchmark
-from torch_frame.gbdt import XGBoost
+from torch_frame.gbdt import CatBoost
 
 parser = argparse.ArgumentParser()
-parser.add_argument('--dataset', type=str, default='california')
+parser.add_argument('--dataset', type=str, default='credit')
 parser.add_argument('--seed', type=int, default=0)
 args = parser.parse_args()
 
@@ -45,9 +45,9 @@ dataset = dataset.shuffle()
 train_dataset, val_dataset, test_dataset = dataset[:0.7], dataset[
     0.7:0.79], dataset[0.79:]
 
-xgb = XGBoost(task_type=TaskType.MULTICLASS_CLASSIFICATION)
-xgb.tune(tf_train=train_dataset.tensor_frame, tf_val=val_dataset.tensor_frame,
-         num_trials=10)
-pred = xgb.predict(tf_test=test_dataset.tensor_frame)
-score = xgb.compute_metric(test_dataset.tensor_frame.y, pred)
+cb = CatBoost(task_type=TaskType.MULTICLASS_CLASSIFICATION)
+cb.tune(tf_train=train_dataset.tensor_frame, tf_val=val_dataset.tensor_frame,
+        num_trials=20)
+pred = cb.predict(tf_test=test_dataset.tensor_frame)
+score = cb.compute_metric(test_dataset.tensor_frame.y, pred)
 print(f'Test acc: {score}')