+++ CHANGELOG.md
@@ -7,9 +7,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
+- Added `CatBoost` baseline and tuned `CatBoost` example. ([#73](https://github.com/pyg-team/pytorch-frame/pull/73))
 - Added `na_strategy` as argument in `StypeEncoder`. ([#69](https://github.com/pyg-team/pytorch-frame/pull/69))
 - Added `NAStrategy` class and impute NaN values in `MutualInformationSort`. ([#68](https://github.com/pyg-team/pytorch-frame/pull/68))
-- Added `XGBoost` baseline and updated extensively tuned `XGBoost` example. ([#57](https://github.com/pyg-team/pytorch-frame/pull/57))
+- Added `XGBoost` baseline and updated tuned `XGBoost` example. ([#57](https://github.com/pyg-team/pytorch-frame/pull/57))
 - Added `CategoricalCatBoostEncoder` and `MutualInformationSort` transforms needed by ExcelFromer ([#52](https://github.com/pyg-team/pytorch-frame/pull/52))
 - Added tutorial example script ([#54](https://github.com/pyg-team/pytorch-frame/pull/54))
 - Added `ResNet` ([#48](https://github.com/pyg-team/pytorch-frame/pull/48))